criado backend curso, modulo, aula, categoria, aulavideo, materialcomplementar

+++ api/src/models/curso/curso.ts
@@ -22,6 +22,7 @@ export const buscarCurso = {
           }
         },
         avaliacoes: true,
+        responsaveltecnico: true,
       },
     });
   },
@@ -36,6 +37,22 @@ export const buscarCursos = {
       where,
       include: {
         categorias: { include: { categoria: true } },
+        modulos: {
+          orderBy: { ordem: 'asc' },
+          include: {
+            avaliacoes: true,
+            aulas: {
+              orderBy: { ordem: 'asc' },
+              include: {
+                avaliacoes: true,
+                materiais: true,
+                videos: true,
+              }
+            }
+          }
+        },
+        avaliacoes: true,
+        responsaveltecnico: true,
       },
       orderBy: { criado_em: 'desc' },
     });
@@ -44,16 +61,28 @@ export const buscarCursos = {
 
 export const criarCurso = {
   async execute(data: any, usuario: any) {
+    const categoriasIds: number[] = data.categorias || [];
+
+    const categoriasValidas = await prisma.categoria.findMany({
+      where: { idCategoria: { in: categoriasIds } }
+    });
+
+    if (categoriasValidas.length === 0) {
+      throw new Error('O curso precisa de pelo menos uma categoria válida');
+    }
+
     const cursoCriado = await prisma.curso.create({
       data: {
         titulo: data.titulo,
         descricao: data.descricao,
         cargaHoraria: data.cargaHoraria,
-        fkEmpresaId: usuario.fkEmpresaId,
+        fkEmpresaId: data.fkEmpresaId,
         fkResponsavelTecnicoId: data.fkResponsavelTecnicoId,
         categorias: {
           createMany: {
-            data: data.categorias.map((id: number) => ({ fkCategoriaId: id }))
+            data: categoriasValidas.map((cat) => ({
+              fkCategoriaId: cat.idCategoria
+            }))
           }
         }
       },
@@ -61,32 +90,57 @@ export const criarCurso = {
         categorias: { include: { categoria: true } }
       }
     });
-    
+
     await registrarEvento({
-        idUsuario: usuario.idUsuario,
-        tipo: "criar",
-        entidade: "curso",
-        entidadeId: cursoCriado.idCurso,
-        descricao: `Curso: ${cursoCriado.titulo} criado com sucesso!`,
-        dadosDepois: cursoCriado,
-      });
+      idUsuario: usuario.idUsuario,
+      tipo: "criar",
+      entidade: "curso",
+      entidadeId: cursoCriado.idCurso,
+      descricao: `Curso: ${cursoCriado.titulo} criado com sucesso!`,
+      dadosDepois: cursoCriado,
+    });
 
     return cursoCriado;
   },
 };
 
 export const editarCurso = {
-  async execute(id: number, data: any) {
+  async execute(id: number, data: any, usuario: any) {
+    const antes = await prisma.curso.findUnique({
+      where: { idCurso: id },
+      include: {
+        categorias: { include: { categoria: true } },
+      }
+    });
+
+    if (!antes) {
+      throw new Error('Nenhum curso encontrado com esse ID.');
+    }
+
+    const categoriasIds: number[] = data.categorias || [];
+
+    const categoriasValidas = await prisma.categoria.findMany({
+      where: { idCategoria: { in: categoriasIds } }
+    });
+
+    if (categoriasValidas.length === 0) {
+      throw new Error('O curso precisa de pelo menos uma categoria válida.');
+    }
+
     await prisma.categoriacurso.deleteMany({ where: { fkCursoId: id } });
+
     const cursoAtualizado = await prisma.curso.update({
       where: { idCurso: id },
       data: {
         titulo: data.titulo,
         descricao: data.descricao,
         cargaHoraria: data.cargaHoraria,
+        fkResponsavelTecnicoId: data.fkResponsavelTecnicoId,
         categorias: {
           createMany: {
-            data: data.categorias.map((id: number) => ({ fkCategoriaId: id }))
+            data: categoriasValidas.map((cat) => ({
+              fkCategoriaId: cat.idCategoria
+            }))
           }
         }
       },
@@ -94,12 +148,46 @@ export const editarCurso = {
         categorias: { include: { categoria: true } }
       }
     });
+
+    await registrarEvento({
+      idUsuario: usuario.idUsuario,
+      tipo: "editar",
+      entidade: "curso",
+      entidadeId: id,
+      descricao: `Curso "${antes.titulo}" atualizado.`,
+      dadosAntes: antes,
+      dadosDepois: cursoAtualizado
+    });
+
     return cursoAtualizado;
-  },
+  }
 };
 
 export const excluirCurso = {
-  async execute(id: number) {
-    await prisma.curso.delete({ where: { idCurso: id } });
+  async execute(id: number, usuario: any) {
+    const curso = await prisma.curso.findUnique({ where: { idCurso: id } });
+
+    if (!curso) {
+      throw new Error('Curso não encontrado ou já foi excluído.');
+    }
+
+    try {
+      await prisma.curso.delete({ where: { idCurso: id } });
+
+      await registrarEvento({
+        idUsuario: usuario.idUsuario,
+        tipo: "excluir",
+        entidade: "curso",
+        entidadeId: curso.idCurso,
+        descricao: `Curso: ${curso.titulo} excluido com sucesso!`,
+      });
+
+    } catch (error: any) {
+      if (error.code === 'P2003') {
+        // Violação de constraint (ex: restrição de chave estrangeira)
+        throw new Error('Não foi possível excluir o curso. Existem registros vinculados.');
+      }
+      throw new Error('Erro ao excluir o curso.');
+    }
   },
 };
